feat: voice skill + wildcard agent proxy + chat rename
Reorganize skills around a thin vestad proxy and an auto-discovered
agent HTTP layer.

**Voice skill**
- New agent/skills/voice/ skill with Deepgram STT (WS relay) and
  ElevenLabs TTS (HTTP streaming) providers
- voice_config.json (keys, voice selection, keyterms, EOT thresholds)
  lives in the container and is managed by the agent via voice_keys.py
  CLI — no master keys on vestad or in the browser
- STT endpoint parameterless; all knobs come from voice_config.json
- Settings UI sends typed `system_message` WS events for voice picks
  and EOT slider changes (silent LLM invocation, no chat bubble)
- Mic button + voice toggle gated on live /voice/status

**Skill auto-discovery**
- Relocated all skills from agent/src/vesta/skills/ to agent/skills/
- Agent scans skills/*/server.py at startup via importlib and mounts
  each skill's routes() as an aiohttp sub-app at /<name>/
- Adding a new skill = drop in server.py; no core/vestad changes

**Vestad collapse**
- Removed dedicated per-agent routes (ws, ws/chat, history, /api/...).
  One wildcard /agents/{name}/{*path} forwards every container-bound
  request. Axum specific-route priority keeps lifecycle routes
  (start/stop/backup/etc.) working.
- Dropped ~500 lines from serve.rs; drops tokio-stream dep, adds reqwest
  for streaming HTTP proxy

**Rename app-chat -> chat**
- URL path /ws/app-chat -> /ws/chat, channel name "app-chat" -> "chat"
- AppChatEvent -> ChatEvent, APP_CHAT_TYPES -> CHAT_TYPES
- MCP tool app_chat_reply -> chat_reply
- components/AppChat/ -> components/Chat/, use-app-chat -> use-chat
- app_chat_reply style guidance moved from skill SKILL.md to MEMORY.md

**Frontend**
- lib/deepgram.ts + lib/elevenlabs.ts -> single lib/voice.ts
- New hooks/use-voice-status.ts (polling, state dedup)
- sendChatEvent exposed as module-level helper on use-chat hook

All tests pass (52/52 agent, ruff+ty clean, cargo+tsc clean).
