+++ agent/skills/default-skills.txt
@@ -1,4 +1,3 @@
-app-chat
 banking
 browser
 dream
