+++ agent/skills/skills-registry/SKILL.md
@@ -30,6 +30,6 @@ ls ~/vesta/skills/
 
 ## Notes
 
-- Skills you install are downloaded from `agent/skills/` in the GitHub repo
+- Skills you install are downloaded from `agent/skills/<name>/` in the GitHub repo
 - Core skills ship pre-installed; optional skills are downloaded on demand
 - After installing a skill that requires setup, read its `SETUP.md`
